Online examinations are the e-examinations conducted through internet(if ,among different organizations) or intranet(if within the organization.Online examination system is a Multiple Choice Questions(MCQ) based examination system that provides a user friendly and less time consuming system to both the student and the instructor. In this system candidate is given a limited time to answer the questions and after the time expiry the answer paper is disabled and score is calculated automatically.The biggest advantage that online examination system introduced is , students can appear the exam from any remote location.It helps in speeding up the process of examination and increases the efficiency of the whole system for which thus approach is acceptable by many exam conducting bodies.
In traditional Paper-based examination, involvement of a lot of paper work
(e.g. printing of exam papers, arranging answer sheets and a lot more manual
works require huge human effort and can also lead to error occurrence. So as
a solution to this “Online Examination System” is introduced through which
we can save a lot of time and get rid of tedious evaluation process as in paperbased examination. It provides an easy user interface to both the students and teachers and the data of students related to their academic, personal or exam related information to any amount of time. Another advantage to thisis as the exam is conducted through the internet, the exam is exposed to the World Wide Web and can be accessed from remote location also